requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
dotenv
alembic
passlib
python-jose
pydantic[email]